# documentation/api.yaml
openapi: 3.0.3
info:
  title: Westco Tack API Course Project Documentation.
  version: 1.0.11
  license:
    name: Apache 2.0
    url: "http://apache.org/"
servers:
  - url: https://educations-apis.herokuapp.com/
    description: Https Hosting Heroku
  - url: http://localhost:5000
    description: Dev Server on Port
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /signUp:
    post:
      summary: Creates a Users
      tags:
        - Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
              example:
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: "12345"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: integer
                  name: string
                  email: string
                  password: string
                  userLessons: array
                  userAdmin: boolen
                  userCourseId: array
              example:
                _id: 62ae948249ef0d92e405b22b
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: "12345"
                userLessons: []
                userAdmin: false
                userCourseId: []
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /signIn:
    post:
      summary: Sign In Users
      tags:
        - Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              example:
                email: Sunnatullox@gmail.com
                password: "12345"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: integer
                  _id: integer
                  name: string
                  email: string
                  password: string
                  userLessons: array
                  userAdmin: boolen
                  userCourseId: array
              example:
                token: TU3ODExNjB9.RoCiHr5wflJbYUWs67NIK4IL3PQcM6fTDg2eeBGsJWo
                _id: 62ae948249ef0d92e405b22b
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$4JUm3qBt.FPGxnqcfYZ08./loDWvYA8VpHxDYdMqVBccieZv/vuBua
                userLessons: []
                userAdmin: false
                userCourseId: []
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...
  /admin/courseCreate:
    post:
      summary: admin Create Course
      tags:
        - Admin Course
      description: Creates Course
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                authorId:
                  type: string
                name:
                  type: string
                price:
                  type: string
                catigory:
                  type: string
                description:
                  type: string
                courseImg:
                  type: string
                  format: binary
                courseLogo:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  adminProfile: string
              example:
                adminProfile: Secsess fully updated User Profile
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /admin/courseUpdate/{id}:
    put:
      summary: User course update
      tags:
        - Admin Course
      description: We add the user to the course and we add the user ID to the course
      consumes:
        - multipart/form-data
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID to updated cours Admin
          example: 62b241cec57596d4147cf96c
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: string
                descrition:
                  type: string
                courseImg:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  userEnroled: string
              example:
                userEnroled: "admin updated cours sucsess fully"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /admin/courseDelete/{id}:
    delete:
      summary: Admin course delete
      tags:
        - Admin Course
      description: admin Delete course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID to Admin deleted course
          example: 62b20de847fc3dfb699ec2a7
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedCourse: string
              example:
                deletedCourse: "Admin sucsess fuly deleted cours"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/courseComment/create:
    post:
      summary: User Course Comment create
      tags:
        - User Course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                courseId:
                  type: string
                userId:
                  type: string
                name:
                  type: string
                description:
                  type: string
              example:
                courseId: 62aec4b4edf4613e87f85cb6
                userId: 62aeb9b09d6840a18ff77f47
                name: Sunnatullox
                description: Kurs Darslari qachon qo'yiladi
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: integer
                  courseId: integer
                  name: string
                  description: string
                  userId: string
                  createdAt: string
                  updatedAt: string
              example:
                "_id": "62b216049920204eb56c309b"
                "userId": "62aeb9b09d6840a18ff77f47"
                "courseId": "62aec4b4edf4613e87f85cb6"
                "name": "Sunnatullox"
                "description": "Kurs Darslari qachon qo'yiladi?"
                "createdAt": "2022-06-21T19:03:32.308Z"
                "updatedAt": "2022-06-21T19:03:32.308Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/courseComment/update/{id}:
    put:
      summary: User course comment update
      tags:
        - User Course
      description: User Course Comment upate
      consumes:
        - multipart/form-data
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the Comment ID to updated comment user
          example: 62b216049920204eb56c309b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              example:
                name: Sunnatullox
                description: Kurs Darslari qachon qo'yiladi
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  _id: integer
                  courseId: integer
                  name: string
                  description: string
                  userId: string
                  createdAt: string
                  updatedAt: string
              example:
                "_id": "62b216049920204eb56c309b"
                "userId": "62aeb9b09d6840a18ff77f47"
                "courseId": "62aec4b4edf4613e87f85cb6"
                "name": "Sunnatullox"
                "description": "Kurs Darslari qachon qo'yiladi?"
                "createdAt": "2022-06-21T19:03:32.308Z"
                "updatedAt": "2022-06-21T19:03:32.308Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/courseComment/delete/{id}:
    delete:
      put:
      summary: User delete course comment
      tags:
        - User Course
      description: user Delete course commet
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID to Admin deleted course
          example: 62b216049920204eb56c309b
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                 deletedCourse: string
              example:
                deletedCourse: "User sucsess fuly deleted cours comment"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...